% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breach_edges}
\alias{breach_edges}
\title{\code{breach_edges} data breach fault tree edge dataset}
\format{
\subsection{\code{breach_edges}}{

A data frame with 7 rows and 2 columns:
\describe{
\item{from}{Unique \code{id} of the source/\code{from} node from which edge originates.}
\item{to}{Unique \code{id} of the destination/\code{to} node that edge connects to.}
}
}
}
\usage{
breach_edges
}
\description{
An example dataset of edges (connections) in the data breach fault tree
described in \link{breach_nodes}.
}
\examples{
data("breach_edges")
head(breach_edges)

}
\seealso{
\link{breach_nodes} for the corresponding nodes dataset
}
\keyword{datasets}
