% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_outcomes_binary}
\alias{db_outcomes_binary}
\title{\code{db_outcomes_binary} database system failure fault tree binary scenarios}
\format{
\subsection{\code{db_outcomes_binary}}{

A tibble with 10 rows and 9 columns:
\describe{
\item{scenario}{Scenario number (1 to 10).}
\item{DC, AF, SF, BF, NF, AUF, HF, MF}{One 0/1 column per basic event; 1 means the event occurred.}
}
}
}
\usage{
db_outcomes_binary
}
\description{
Ten example scenarios for the basic events of the database system failure fault tree, coded
1 if the event occurred and 0 if not. Use it with \code{\link[=quantify]{quantify()}} to evaluate
whether the top event occurs under each scenario, or with \code{\link[=populate]{populate()}}.
}
\examples{
data("db_outcomes_binary")
head(db_outcomes_binary)

}
\seealso{
\link{db_probs}, \link{db_outcomes_prob}, \code{\link[=quantify]{quantify()}}, \code{\link[=populate]{populate()}}
}
\keyword{datasets}
