% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_outcomes_rates}
\alias{db_outcomes_rates}
\title{\code{db_outcomes_rates} database system failure fault tree failure rates}
\format{
\subsection{\code{db_outcomes_rates}}{

A tibble with 8 rows and 3 columns:
\describe{
\item{event}{Name of the basic event: \code{"DC"} = Data corruption, \code{"AF"} = Access failure, \code{"SF"} = Storage failure, \code{"BF"} = Backup failure, \code{"NF"} = Network failure, \code{"AUF"} = Authentication failure, \code{"HF"} = Hardware failure, \code{"MF"} = Monitoring failure.}
\item{lambda}{Failure rate (failures per hour).}
\item{time_unit}{Unit of time for \code{lambda}: \code{"hours"}.}
}
}
}
\usage{
db_outcomes_rates
}
\description{
Example exponential failure rates for the basic events of the database system failure fault
tree, expressed per hour.
}
\examples{
data("db_outcomes_rates")
db_outcomes_rates

}
\seealso{
\link{db_probs}, \link{db_outcomes_prob}
}
\keyword{datasets}
