% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_probs}
\alias{db_probs}
\title{\code{db_probs} database system failure fault tree event probabilities}
\format{
\subsection{\code{db_probs}}{

A tibble with 8 rows and 2 columns:
\describe{
\item{event}{Name of the basic event: \code{"DC"} = Data corruption, \code{"AF"} = Access failure, \code{"SF"} = Storage failure, \code{"BF"} = Backup failure, \code{"NF"} = Network failure, \code{"AUF"} = Authentication failure, \code{"HF"} = Hardware failure, \code{"MF"} = Monitoring failure.}
\item{probability}{Probability that the event occurs, between 0 and 1.}
}
}
}
\usage{
db_probs
}
\description{
Example failure probabilities for the basic events of the database system failure fault tree
(\code{db_nodes}, \code{db_edges}). Pass it as \code{event_probs} to \code{\link[=populate]{populate()}}.
}
\examples{
data("db_probs")
db_probs

}
\seealso{
\link{db_outcomes_prob}, \link{db_outcomes_binary}, \code{\link[=populate]{populate()}}
}
\keyword{datasets}
