% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakeedges.R
\docType{data}
\name{fakeedges}
\alias{fakeedges}
\title{\code{fakeedges} fault tree edge dataset}
\format{
\subsection{\code{fakeedges}}{

A data frame with 11 rows and 2 columns:
\describe{
\item{from}{Unique \code{id} of the source/\code{from} node from which edge originates.}
\item{to}{unique \code{id} of the destination/\code{to} node that edge connects to.}
...
}
}
}
\usage{
fakeedges
}
\description{
An example dataset of edges (connections) in a demonstration fault tree.
}
\keyword{datasets}
