% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakenodes.R
\docType{data}
\name{fakenodes}
\alias{fakenodes}
\title{\code{fakenodes} fault tree nodes dataset}
\format{
\subsection{\code{fakenodes}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{id}{Unique identifer (1 to N) for each node. Some events occur multiple times, but each time they have a unique \code{id}}
\item{event}{name of event. \code{"T"} mean top event. \code{"G1"}, \code{"G2"}, etc. mean gates.}
\item{type}{\code{factor} classification as "top", "and", "or", or "not" (meaning "not" a gate).}
...
}
}
}
\usage{
fakenodes
}
\description{
An example dataset of nodes in a demonstration fault tree.
}
\keyword{datasets}
