% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_security_edges.R
\docType{data}
\name{it_security_edges}
\alias{it_security_edges}
\title{\code{it_security_edges} IT security (data leak) fault tree edge dataset}
\format{
\subsection{\code{it_security_edges}}{

A data frame with 15 rows and 2 columns:
\describe{
\item{from}{Source node \code{id}.}
\item{to}{Destination node \code{id}.}
}
}
}
\usage{
it_security_edges
}
\description{
An example dataset of edges (connections) in an IT security fault tree
for sensitive data leaving the organization.
}
\examples{
data("it_security_edges")
head(it_security_edges)
dim(it_security_edges)

}
\seealso{
\code{\link[tidyfault]{it_security_nodes}} for the corresponding nodes dataset
}
\keyword{datasets}
