% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{security_outcomes_rates}
\alias{security_outcomes_rates}
\title{\code{security_outcomes_rates} security breach fault tree failure rates}
\format{
\subsection{\code{security_outcomes_rates}}{

A tibble with 7 rows and 3 columns:
\describe{
\item{event}{Name of the basic event: \code{"VE"} = Vulnerability exists, \code{"ES"} = Exploit successful, \code{"PH"} = Phishing, \code{"MW"} = Malware, \code{"UA"} = Unauthorized access, \code{"WP"} = Weak password, \code{"N2F"} = No 2FA.}
\item{lambda}{Failure rate (failures per day).}
\item{time_unit}{Unit of time for \code{lambda}: \code{"days"}.}
}
}
}
\usage{
security_outcomes_rates
}
\description{
Example exponential failure rates for the basic events of the security breach fault
tree, expressed per day.
}
\examples{
data("security_outcomes_rates")
security_outcomes_rates

}
\seealso{
\link{security_probs}, \link{security_outcomes_prob}
}
\keyword{datasets}
