% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_edges.R
\docType{data}
\name{ai_edges}
\alias{ai_edges}
\title{\code{ai_edges} AI agent failure fault tree edge dataset}
\format{
\subsection{\code{ai_edges}}{

A data frame with 5 rows and 2 columns:
\describe{
\item{from}{Unique \code{id} of the source/\code{from} node from which edge originates.}
\item{to}{Unique \code{id} of the destination/\code{to} node that edge connects to.}
}
}
}
\usage{
ai_edges
}
\description{
An example dataset of edges (connections) in an AI agent failure fault tree.
}
\examples{
data("ai_edges")
head(ai_edges)
dim(ai_edges)

}
\seealso{
\code{\link[tidyfault]{ai_nodes}} for the corresponding nodes dataset
}
\keyword{datasets}
