% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{calculate() Function}
\usage{
calculate(f)
}
\arguments{
\item{data}{(Required) function outputted by \code{equate()}, including arguments for all input events in the fault tree (excluing gates or top events)}
}
\description{
This function takes a supplied function \code{f} and computes the system failure outcomes (1 = failure vs. 0 = not) given every possible combination of input events failing (1) or not failing (0). Outputs a \code{data.frame}.
}
\examples{

# Load dependencies
library(tidyverse)
library(tidyfault)
library(QCA)

# Load example data into our environment
data("fakenodes")
data("fakeedges")

# Extract minimum cutset from fault tree data
curate(nodes = fakenodes, edges = fakeedges) \%>\%
   equate() \%>\%
   formulate() \%>\%
   calculate() \%>\%
   concentrate() \%>\% 
   tabulate()
}
\keyword{boolean}
\keyword{equation}
\keyword{fault}
\keyword{logic}
\keyword{tree}
