% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrate.R
\name{concentrate}
\alias{concentrate}
\title{concentrate() Function}
\usage{
concentrate(data, method = "mocus")
}
\arguments{
\item{data}{(Required) data.frame containing output from \code{curate()} function for the \code{"mocus"} algorithm. (Or output from \code{calculate()} for the \verb{'CCubes' algorithm}.}

\item{method}{(Optional) By default, runs \code{"mocus"} algorithm on output from \code{curate()} and simplifies to minimum cutsets. Alternatively, can run \code{"CCubes"} algorithm on the output from \code{calculate()} (but best on smaller datasets).}
}
\description{
This function \emph{concentrates} the boolean equation of a fault tree to find the minimum cutsets in the fault tree. Takes a data.frame outputted by the \code{curate()} function and applies the \code{mocus()} algorithm, or takes a data.frame outputted by the \code{calculate()} function and applies the QCA package's \code{"CCubes"} algorithm to perform boolean minimization.
}
\examples{

# Load dependencies
library(tidyverse)
library(tidyfault)
library(QCA)

# Load example data into our environment
data("fakenodes")
data("fakeedges")

# Extract minimum cutset from fault tree data
curate(nodes = fakenodes, edges = fakeedges) \%>\%
   concentrate() \%>\% 
   tabulate()
}
\keyword{cutset}
\keyword{fault}
\keyword{minimalization}
\keyword{minimum}
\keyword{qca}
\keyword{tree}
