% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrate.R
\name{concentrate}
\alias{concentrate}
\title{concentrate() Function}
\usage{
concentrate(data)
}
\arguments{
\item{data}{(Required) data.frame containing truth table of N possible combinations of input events (eg. \code{A}, \code{B}, \code{C}) as either failing (1) or not failing (0), and the corresponding \code{outcome} measured by \code{calculate()} showing overall system failure (1) or not (0). \code{outcome} must be the last column.}
}
\description{
This function \emph{concentrates} a \code{data.frame} of N possible combinations of input events in the graph (not counting gates or top events), by applying boolean minimalization to find the minimum cutsets in the fault tree. Uses the \code{QCA} package and the \code{"CCubes"} algorithm to perform boolean minimization.
}
\examples{

# Load dependencies
library(tidyverse)
library(tidyfault)
library(QCA)

# Load example data into our environment
data("fakenodes")
data("fakeedges")

# Extract minimum cutset from fault tree data
curate(nodes = fakenodes, edges = fakeedges) \%>\%
   equate() \%>\%
   formulate() \%>\%
   calculate() \%>\%
   concentrate() \%>\% 
   tabulate()
}
\keyword{cutset}
\keyword{fault}
\keyword{minimalization}
\keyword{minimum}
\keyword{qca}
\keyword{tree}
