% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{curate}
\alias{curate}
\title{curate() Function}
\usage{
curate(nodes, edges)
}
\arguments{
\item{nodes}{(Required) data.frame of N \code{nodes} (1 to N rows), with a unique \code{id} for each node, an \code{event} name describing the node (some events occur multiple times), and a \code{type} vector categorizing nodes as \code{"and"} | \code{"or"} gates, \code{"top"} nodes, or \code{"not"} a gate or top node (ie. a normal node)}

\item{edges}{(Required) data.frame of N \code{edges} (1 to N rows), showing ties between the \code{from} node's \code{id} to the \code{to} node's \code{id}.}
}
\description{
This function \emph{curates} a \code{data.frame} of N \code{gates} and the \code{sets} of events immediately linked to those gates.
}
\examples{

# Load dependencies
library(tidyverse)
library(tidyfault)
library(QCA)

# Load example data into our environment
data("fakenodes")
data("fakeedges")

# Extract minimum cutset from fault tree data
curate(nodes = fakenodes, edges = fakeedges) \%>\%
   equate() \%>\%
   formulate() \%>\%
   calculate() \%>\%
   concentrate() \%>\% 
   tabulate()
}
\keyword{fault}
\keyword{tree}
