% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_edges.R
\docType{data}
\name{db_edges}
\alias{db_edges}
\title{\code{db_edges} database system failure fault tree edge dataset}
\format{
\subsection{\code{db_edges}}{

A data frame with 12 rows and 2 columns:
\describe{
\item{from}{Unique \code{id} of the source/\code{from} node from which edge originates.}
\item{to}{Unique \code{id} of the destination/\code{to} node that edge connects to.}
}
}
}
\usage{
db_edges
}
\description{
An example dataset of edges (connections) in a database system failure fault tree.
}
\examples{
data("db_edges")
head(db_edges)
dim(db_edges)

}
\seealso{
\code{\link[tidyfault]{db_nodes}} for the corresponding nodes dataset
}
\keyword{datasets}
