% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate.R
\name{equate}
\alias{equate}
\title{equate() Function}
\usage{
equate(data)
}
\arguments{
\item{data}{(Required) data.frame of N gates, containing columns for \code{gate}, \code{type}, and \code{set}. Outputted by \code{curate()}.}
}
\description{
This function extracts the equation for a fault tree from a \code{data.frame} of \code{gate}s and \code{set}s, outputted by \code{curate()}.
}
\examples{

# Load dependencies
library(tidyverse)
library(tidyfault)
library(QCA)

# Load example data into our environment
data("fakenodes")
data("fakeedges")

# Extract minimum cutset from fault tree data
curate(nodes = fakenodes, edges = fakeedges) \%>\%
   equate() \%>\%
   formulate() \%>\%
   calculate() \%>\%
   concentrate() \%>\% 
   tabulate()
}
\keyword{equation}
\keyword{fault}
\keyword{tree}
