% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulate.R
\name{formulate}
\alias{formulate}
\title{formulate() Function}
\usage{
formulate(formula)
}
\arguments{
\item{formula}{(Required) a character string listing the boolean logic equation of a fault tree.}
}
\description{
This function \emph{formulates} a \code{function} that can compute probabilities of system failure. To do so, it converts a character string describing the boolean equation of a fault tree into a \code{function} that can compute probabilities of system failure into that function. Handles AND and OR operations.
}
\examples{

# Load dependencies
library(tidyverse)
library(tidyfault)
library(QCA)

# Load example data into our environment
data("fakenodes")
data("fakeedges")

# Extract minimum cutset from fault tree data
curate(nodes = fakenodes, edges = fakeedges) \%>\%
   equate() \%>\%
   formulate() \%>\%
   calculate() \%>\%
   concentrate() \%>\% 
   tabulate()
}
\keyword{boolean}
\keyword{equation}
\keyword{fault}
\keyword{formula}
\keyword{tree}
