% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate.R
\name{gate}
\alias{gate}
\title{gate() Function}
\usage{
gate(data, group = "id", gate = "type", size = 1, res = 50)
}
\arguments{
\item{data}{(Required) data.frame of any nodes. Must include at least 1 row where gate vector = "and", "or", or "top"}

\item{gate}{(Default = \code{"type"}) string name of vector containing gate types. Default is "type"}

\item{size}{(Default = 1) Default diameter of gate polygons.}

\item{res}{(Default = 50) Default number of line segments spanning x axis.}

\item{id}{(Default = \code{"id"}) string name of unique id vector in \code{data}.}
}
\description{
This internal function gathers all gate polygons for a set of supplied gate types and their x and y coordinates.
}
\keyword{fault}
\keyword{gate}
\keyword{maker}
\keyword{polygon}
\keyword{tree}
