% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_and.R
\name{gate_and}
\alias{gate_and}
\title{gate_and() Function}
\usage{
gate_and(size = 2, res)
}
\arguments{
\item{res}{(Required) A numeric value describing the resolution, or number of line segments spanning x-axis.}
}
\description{
This function produces the top gate polygons for fault trees.
}
\keyword{fault}
\keyword{gate}
\keyword{maker}
\keyword{polygon}
\keyword{tree}
