% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_or.R
\name{gate_or}
\alias{gate_or}
\title{gate_or() Function}
\usage{
gate_or(size = 2, res)
}
\arguments{
\item{res}{(Required) A numeric value describing the resolution, or number of line segments spanning x-axis.}
}
\description{
This function produces the OR gate polygons for fault trees.
}
\keyword{fault}
\keyword{gate}
\keyword{maker}
\keyword{polygon}
\keyword{tree}
