% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gate.R
\name{get_gate}
\alias{get_gate}
\title{get_gate() Function}
\usage{
get_gate(x, y, gate, size = 1, res = 50)
}
\arguments{
\item{x}{(Required) x coordinate of gate in the layout.}

\item{y}{(Required) y coordinate of gate in the layout.}

\item{gate}{(Required) vector of gate names. Must be one of c("top", "and", "or")}
}
\description{
This function produces the gate polygons for a set of supplied gate types and their x and y coordinates.
}
\keyword{fault}
\keyword{gate}
\keyword{maker}
\keyword{polygon}
\keyword{tree}
