% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illustrate.R
\name{illustrate}
\alias{illustrate}
\title{illustrate() Function}
\usage{
illustrate(
  nodes,
  edges,
  type = c("nodes", "edges", "both", "all"),
  node_key = "id",
  layout = "tree"
)
}
\arguments{
\item{nodes}{(Required) dataset of nodes, including an \code{id} field (or whatever is written in \code{node_key})}

\item{edges}{(Required) dataset of edges, including \code{from} and \code{to} fields corresponding to the \code{id} of each node}

\item{type}{(Optional) By default, returns \code{"both"} the \code{"nodes"} and \code{"edges"} as items in a list. Alternatively, you can select just \code{"nodes"}, \code{"edges"}, or \code{"all"} to receive other formats of outputs.}

\item{node_key}{(Optional) By default, \code{node_key} = \verb{"id'} from \code{nodes}}

\item{layout}{(Optional) By default, network layout is \code{"tree"}. \code{"dendrogram"} also works pretty well. See \code{ggraph} package for more layout options.}
}
\description{
This function takes a dataset of nodes and edges dataset from a fault tree, calculate a \strong{tree} layout using \code{tidygraph} and \code{ggraph}, and then outputs ggplot-friendly data.frames of nodes and edges.
}
\keyword{ggplot}
\keyword{network}
\keyword{visualize}
