% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocus.R
\name{mocus}
\alias{mocus}
\title{mocus() Function}
\usage{
mocus(data)
}
\arguments{
\item{data}{(Required) data.frame containing gates and their sets, outputted by \code{curate()}.}
}
\description{
This function \emph{simplifies} a \code{data.frame} of gates and their sets identified by \code{curate()}, generating each cutset in the fault tree using a simple implementation of the \code{mocus()} algorithm.
}
\keyword{boolean}
\keyword{cutset}
\keyword{fault}
\keyword{tree}
