% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_edges.R
\docType{data}
\name{security_edges}
\alias{security_edges}
\title{\code{security_edges} security breach fault tree edge dataset}
\format{
\subsection{\code{security_edges}}{

A data frame with 11 rows and 2 columns:
\describe{
\item{from}{Unique \code{id} of the source/\code{from} node from which edge originates.}
\item{to}{Unique \code{id} of the destination/\code{to} node that edge connects to.}
}
}
}
\usage{
security_edges
}
\description{
An example dataset of edges (connections) in a security breach fault tree.
}
\seealso{
\code{\link[tidyfault]{security_nodes}} for the corresponding nodes dataset
}
\keyword{datasets}
