% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfault-package.R
\docType{package}
\name{tidyfault-package}
\alias{tidyfault}
\alias{tidyfault-package}
\title{tidyfault: Fault Tree Analysis in R with Tidyverse Coding}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R Package for tidy fault tree analysis! Uses tidygraph and tidyverse packages, among others, to visualize fault trees, identify minimal cutsets, and other quantities of interest.
}
\author{
\strong{Maintainer}: Timothy Fraser \email{timothy.fraser.1@gmail.com} (\href{https://orcid.org/0000-0002-4509-0244}{ORCID})

Authors:
\itemize{
  \item Jingyao Tong \email{jingyaotong6@gmail.com} (\href{https://orcid.org/0009-0009-6147-756X}{ORCID})
}

}
\keyword{internal}
